+++ carbon/include/flipper/uart0.h
@@ -13,7 +13,7 @@ void uart0_put(uint8_t byte);
 uint8_t uart0_get(void);
 void uart0_enable(void);
 void uart0_disable(void);
-int uart0_write(void *src, size_t length);
-int uart0_read(void *dst, size_t length);
+int uart0_write(void *src, uint32_t length);
+int uart0_read(void *dst, uint32_t length);
 
 #endif
