+++ runtime/src/device.c
@@ -1,8 +1,8 @@
 #include <flipper.h>
 
 /* Creates a new libflipper device. */
-struct _lf_device *lf_device_create(int (* read)(struct _lf_device *device, void *dst, size_t length),
-									int (* write)(struct _lf_device *device, void *src, size_t length),
+struct _lf_device *lf_device_create(int (* read)(struct _lf_device *device, void *dst, uint32_t length),
+									int (* write)(struct _lf_device *device, void *src, uint32_t length),
 									int (* release)(struct _lf_device *device)) {
 	struct _lf_device *device = NULL;
 	device = (struct _lf_device *)calloc(1, sizeof(struct _lf_device));
