+++ utils/fvm/src/api/usart.c
@@ -10,12 +10,12 @@ LF_FUNC("usart") int usart_ready(void) {
 	return lf_success;
 }
 
-LF_FUNC("usart") int usart_push(void *src, size_t length) {
+LF_FUNC("usart") int usart_push(void *src, uint32_t length) {
 	printf("Pushing '%s' to the usart bus.\n", src);
 	return lf_success;
 }
 
-LF_FUNC("usart") int usart_pull(void *dst, size_t length) {
+LF_FUNC("usart") int usart_pull(void *dst, uint32_t length) {
 	printf("Pulling from the usart bus.\n");
 	return lf_success;
 }
